+++ tests/testthat/test_citations.R
@@ -21,10 +21,13 @@ test_that("layer_citations returns expected citations", {
 })
 
 test_that("layer_citations returns layer specific citations", {
-  testthat::skip_on_travis()
+  testthat::skip_on_ci()
   testthat::skip_on_cran()
   
-  if(length(sdmpredictors:::get_sysdata()$bibentries$lnk_bibentry$layers) > 0) {
+  lyr_cit <- sdmpredictors:::get_sysdata()$bibentries$lnk_bibentry$layers
+  expect_null(lyr_cit)
+
+  if(length(lyr_cit) > 0) {
     fail(message = "IMPLEMENT THIS TEST when there are layer specific bibentries added to bibentry")
   }
 })
@@ -37,14 +40,15 @@ test_that("each dataset has at least one citation", {
   }
 })
 
-test_that("each layer has at least one citation", {
-  testthat::skip_on_cran()
-  layers <- list_layers()
-  for(i in 1:nrow(layers)) {
-    ref <- layer_citations(layers[i,])
-    testthat::expect_gt(length(ref), 0)
-  }
-})
+# Run only once, no need to test more than 1000 layers everytime
+# test_that("each layer has at least one citation", {
+#   testthat::skip_on_cran()
+#   layers <- list_layers()
+#   for(i in 1:nrow(layers)) {
+#     ref <- layer_citations(layers[i,])
+#     testthat::expect_gt(length(ref), 0)
+#   }
+# })
 
 test_that("citations are returned", {
   lcitations <- layer_citations()