+++ NEWS.md
@@ -1,3 +1,9 @@
+# sdmpredictors 0.2.10
+
+Behrmann equal areas projection is now created on the fly from the original WGS84 layer
+Bio-Oracle layers are downloaded from bio-oracle.org instead of lifewatch.be
+sdmpredictors downloads layers now from specific URLs instead of relying on lifewatch.be/sdmpredictors and the layer_code
+
 # sdmpredictors 0.2.9
 
 Add bio-oracle v2.1